backend/requirements.txt
langgraph
langchain-google-genai
langchain-community
langchain-core
tavily-python
python-dotenv
fastapi
uvicorn